Update Problem:2-Identity_vowels.cs
Added some modifications to the code according to the rules mentioned in moodle.

Added a do while loop to give the user access to continue or quit the program.

the program stores all the vowels in an array and checks whether the entered alphabet exist in the array. if it exists in the array the bool variable flag is set to true. According to the value of bool the output is displayed.

The program uses array and for loop rather than using if/else condition or switch statements for reducing the number of line of code.